onchain/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
web3==6.20.0
python-dotenv==1.0.1
py-solc-x==2.0.3
pydantic==2.9.2
matplotlib==3.9.2
